Add files via upload
in this program, there are two functions. In the first function, with the help of state variable, computer can know whether the guessed word is too high or too low.
this is a step-by-step process.
whereas, in the second method, the output is the number of takes that the computer takes to guess the word. it is not a step by step procedure.